--- src/main/kotlin/uk/gov/justice/digital/hmpps/approvedpremisesapi/model/ValidatableActionResult.kt
@@ -1,7 +1,0 @@
-package uk.gov.justice.digital.hmpps.approvedpremisesapi.model
-
-sealed interface ValidatableActionResult<EntityType> {
-  data class Success<EntityType>(val entity: EntityType) : ValidatableActionResult<EntityType>
-  data class FieldValidationError<EntityType>(val validationMessages: ValidationErrors) : ValidatableActionResult<EntityType>
-  data class GeneralValidationError<EntityType>(val message: String) : ValidatableActionResult<EntityType>
-}
